requirements.txt
crewai
langchain_groq
streamlit
python-dotenv
exa-py
pandas
litellm
langchain_community